UBUNTU: [Config] Use zImage for installing vmlinux on ppc32/ppc64
This does not apply to e500* kernels. The reason for doing this is because
CONFIG_DEBUG_INFO=y is being forced now, which means using raw vmlinux
was giving a >140Meg kernel. Instead of mucking around in the machinery,
just use the compressed zImage wrapper.

CAVEAT: We aren't sure at this poing that it will work on older style
boot loaders like quik and BootX.

Signed-off-by: Ben Collins <ben.c@servergy.com>